services/inference/requirements.txt
grpcio==1.64.1
protobuf==4.25.3
redis==5.0.4
anthropic==0.97.0
google-genai==1.47.0
openai==2.32.0
pydantic-settings==2.11.0
